Vendor update (#23)
* time resuming panel bug fixed

* chest, inventory and healthbar asset update + chest coin system improvement

* item dropdown

* equippable slots + equipment update

* vendor script / inventory

* karma system start

keeps track of kills and steals, increases prices

* item fix

* minor fixes

+++ Assets/Scripts/Enemy/EnemyAttack.cs
@@ -7,7 +7,7 @@ public class EnemyAttack : MonoBehaviour
     [SerializeField] private float attackRange = 2f;
     [SerializeField] private float attackAngle = 60f;
     [SerializeField] private float attackCooldown = 1f;
-    [SerializeField] private int attackDamage = 15;
+    [SerializeField] private float attackDamage = 15;
 
     [SerializeField]
     private Transform hitEffect;
